django_face_recognition/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
